+++ tests/conftest.py
@@ -2,3 +2,5 @@ import warnings
 
 # Turn deprecation warnings into errors
 warnings.simplefilter('error', DeprecationWarning)
+
+pytest_plugins = ['arctic.fixtures.arctic']